be/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
flask_sqlalchemy==3.1.1
numpy==2.0.0
Pillow==10.3.0
SQLAlchemy==2.0.31
thread==2.0.3
flask_bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
requests==2.32.3